refactor : domain 패키지 분리: Course 패키지로 이동
